requirements.pip.txt
aiohttp == 3.6.2
appdirs == 1.4.3
cryptography == 2.8
docutils == 0.15.2
doit == 0.31.1
flake8 == 3.7.9
jsonpatch == 1.23
jsonschema == 3.0.2
mako == 1.1.0
packaging == 19.2
psutil == 5.6.5
pyserial == 3.4
pytest == 5.0.1
pytest-asyncio == 0.10.0
pytest-cov == 2.7.1
pytest-repeat == 0.8.0
pytest-timeout == 1.3.3
PyYAML == 5.1.2
setuptools == 41.0.1
Sphinx == 2.2.1
wheel == 0.33.4
